+++ docker-compose.yml
@@ -1,0 +1,15 @@
+version: '3.4'
+
+services:
+  sentimentanalysispipeline:
+    image: sentimentanalysispipeline
+    build:
+      context: .
+      dockerfile: ./Dockerfile
+    volumes:
+      - sentiment-models:/app/models  # Ajout du volume ici
+    ports:
+      - "5000:5000"  # ajoute si tu exposes une API, sinon tu peux ignorer
+
+volumes:
+  sentiment-models:  # Définition explicite du volume
