+++ requirements.txt
@@ -1,5 +1,7 @@
 numpy
+pandas
 seaborn
 scikit-learn
-transformers
 torch
+transformers
+tkinter
